fix: 完成ads 小游戏数据统计

+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-ads/src/main/java/com/leo/ad/codriver/ads/entity/AdsDailyLabOetaBaseReport.java
@@ -13,7 +13,7 @@ import lombok.Data;
 public class AdsDailyLabOetaBaseReport implements BaseEntity {
 
     private Long id;
-    private Long dates;
+    private Integer dates;
     private String version;
     private String pkg;
     private Long userType;
